Clarifying build steps on Windows
FIX: add <QAction> in UI/frontend-plugins/frontend-tools/scripts.cpp
A compilation error was returned without #include <QAction>

FIX: closes #74
It is better to keep CMake that way :
Windows user of the CMake script should be careful on which OpenSSL is used.
Passing the variable OPENSSL_ROOT_DIR to the install location of OpenSSL makes sure of that.