Update README.md
No support for Janus will be provided by default. Bruden of proof (that there is a bug on the OS Side) is on the one(s) opening the tickets.
There is no commercial support for this fork of OBS.
Janus support will only be done through the WHIP protocol.
Commercial uspport for Janus server side is available [here](https://janus.conf.meetecho.com/support).